refactor(invoice): cleanup task and workflow module
